Get rid of app cache
The main activity needs post link to be opened in desktop view and the
view post link activity needs to be opened in mobile view. This messes
up the cache use.

Also the cache size is pretty huge considering the fact that we have
many links open here.

So bye bye cache !! It was nice working with you

:D

Signed-off-by: Ashutosh Sathe <2ashutoshbs@gmail.com>